Create new test_functions module and consolidate run_func tests
