+++ pymatbridge/tests/test_run_code.py
@@ -64,33 +64,3 @@ class TestRunCode:
             npt.assert_equal(message, "'this_is_nonsense' undefined near line 1 column 1")
         else:
             npt.assert_equal(message, "Undefined function or variable 'this_is_nonsense'.")
-
-
-    def test_nargout(self):
-        res  = self.mlab.run_func('svd', np.array([[1,2],[1,3]]), nargout=3)
-        U, S, V = res['result']
-        npt.assert_almost_equal(U, np.array([[-0.57604844, -0.81741556],
-                                             [-0.81741556, 0.57604844]]))
-
-        npt.assert_almost_equal(S, np.array([[ 3.86432845, 0.],
-                                             [ 0., 0.25877718]]))
-
-        npt.assert_almost_equal(V, np.array([[-0.36059668, -0.93272184],
-                                             [-0.93272184, 0.36059668]]))
-
-        res = self.mlab.run_func('svd', np.array([[1,2],[1,3]]), nargout=1)
-        s = res['result']
-        npt.assert_almost_equal(s, [[ 3.86432845], [ 0.25877718]])
-
-        res = self.mlab.run_func('close', 'all', nargout=0)
-        assert res['result'] == []
-
-    def test_tuple_args(self):
-        res = self.mlab.run_func('ones', (1, 2))
-        npt.assert_almost_equal(res['result'], [[1, 1]])
-
-        res = self.mlab.run_func('chol',
-                                 (np.array([[2, 2], [1, 1]]), 'lower'))
-        npt.assert_almost_equal(res['result'],
-                                [[1.41421356, 0.],
-                                 [0.70710678, 0.70710678]])
